Upgrade to Spring Boot 4.0.5, Spring Cloud 2025.1.1, resolve OSS vulnerabilities
- Bump spring-boot-starter-parent 3.5.13 → 4.0.5 (Spring Framework 7.0.6, Hibernate 7.2.7)
- Bump spring-cloud-dependencies 2025.0.2 → 2025.1.1
- Rename gateway dependency: spring-cloud-starter-gateway → spring-cloud-starter-gateway-server-webflux
- Update gateway routes config key to spring.cloud.gateway.server.webflux.*
- Remove pinned postgresql/h2 versions in liquibase module (now BOM-managed)
- Fix AccountDao generic type Long → Integer to match Account @Id field
- Add @ActiveProfiles("test") to CustomerApplicationTest

+++ account/src/main/java/com/halversondm/cloud/AccountDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AccountDao extends CrudRepository<Account, Long> {
+public interface AccountDao extends CrudRepository<Account, Integer> {
 
     Account findByNumber(String number);
